pseudo modularizing elliptic
