+++ include/eve/module/elliptic/regular/impl/ellint_rj.hpp
@@ -7,30 +7,13 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/detail/has_abi.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/constant/eps.hpp>
-#include <eve/constant/half.hpp>
-#include <eve/constant/one.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/average.hpp>
-#include <eve/function/ellint_rc.hpp>
-#include <eve/function/ellint_rd.hpp>
-#include <eve/function/ellint_rf.hpp>
-#include <eve/function/fma.hpp>
-#include <eve/function/fms.hpp>
-#include <eve/function/if_else.hpp>
-#include <eve/function/is_nltz.hpp>
-#include <eve/function/is_nez.hpp>
-#include <eve/function/max.hpp>
-#include <eve/function/rec.hpp>
-#include <eve/function/rsqrt.hpp>
-#include <eve/function/sqrt.hpp>
-#include <eve/function/pow_abs.hpp>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/ieee.hpp>
 #include <eve/detail/hz_device.hpp>
-
+#include <eve/module/elliptic/regular/ellint_rc.hpp>
+#include <eve/module/elliptic/regular/ellint_rd.hpp>
+#include <eve/module/elliptic/regular/ellint_rf.hpp>
 
 namespace eve::detail
 {
